@@ -6,8 +6,8 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -38,7 +38,7 @@ import java.util.concurrent.TimeUnit;
  */
 public class SensorFragment extends Fragment {
 
-    String TAG = "SensorFrag";
+    final static String TAG = "SensorFrag";
     TextView logger;
     static final int REQUEST_OAUTH = 1;
     OnDataPointListener mlistener;
